updated: benchmark_collect.sh
Now keeps results from last program execution.

+++ code/cpp_version/benchmark_collect.sh
@@ -3,51 +3,52 @@
 
 declare -a nParticles=(10 100 1000 10000)
 
-#pos=$(( ${#nParticles[*]} - 1 ))
-#last=${nParticles[$pos]}
+pos=$(( ${#nParticles[*]} - 1 ))
+last=${nParticles[$pos]}
 
 filename_serial="benchmark_serial.txt"
 
 echo "num_particles" "," "time_ns" > $filename_serial
 
 # TODO: Add up average times.
-
 for i in "${nParticles[@]}"
 do
-  iParticles=$i
-  # echo "nParticles="$i >> $filename_serial
-  # TODO: fix so that avg_duration is added to line that says nParticles=...
   avg_time=`./particles_serial n=$i delta=0.5 total_time_interval=100 | cut -d= -f2`
 
-  echo $iParticles "," $avg_time >> $filename_serial
+  echo $i "," $avg_time >> $filename_serial
 
-  rm particle_positions/positions_*.vtk
-
-  # mpirun -np $rank ./main $i $i 100 0.01 > results/res_r_"$rank"_n_"$num_threads"_"$i".txt
+  # Only remove particle positions if not in last program execution.
+  # I.e., keep only particle positions from last program execution.
+  if [ $i -ne $last ]
+  then
+    rm particle_positions/positions_*.vtk
+  fi
 done
 
 # TODO: Execute code using OpenACC.
-# declare -a nParticles=(10 100 1000 10000)
-
-#pos=$(( ${#nParticles[*]} - 1 ))
-#last=${nParticles[$pos]}
-
+# Keeping the following three commented in case we later want to execute
+# parallel code with more particles than in serial version. But we'd have
+# to have an nParticles2 variable or something of the sort.
+# declare -a nParticles2=(10 100 1000 10000)
+# pos=$(( ${#nParticles2[*]} - 1 ))
+# #last=${nParticles2[$pos]}
 filename_parallel="benchmark_parallel.txt"
 
 echo "num_particles" "," "time_ns" > $filename_parallel
 
 # TODO: Add up average times.
-
 for i in "${nParticles[@]}"
 do
-  iParticles=$i
-  # echo "nParticles="$i >> $filename_serial
-  # TODO: fix so that avg_duration is added to line that says nParticles=...
   avg_time=`./particles_parallel n=$i delta=0.5 total_time_interval=100 | cut -d= -f2`
 
-  echo $iParticles "," $avg_time >> $filename_parallel
+  echo $i "," $avg_time >> $filename_parallel
 
-  rm particle_positions/positions_*.vtk
+  # Only remove particle positions if not in last program execution.
+  # I.e., keep only particle positions from last program execution.
+  if [ $i -ne $last ]
+  then
+    rm particle_positions/positions_*.vtk
+  fi
 done
 
 echo All done
